all behaviours added

+++ city.cpp
@@ -8,12 +8,12 @@
 
 using namespace std;
 
-city::city(std::string name, double xCoord, double yCoord) : name(name) {
-    if (xCoord>1000.0 || xCoord<0.0) {
+city::city(std::string name, int xCoord, int yCoord) : name(name) {
+    if (xCoord>1000 || xCoord<0) {
         cerr << "Please make sure X-Coordinate is between 0.0 and 1000.0";
         throw "CoordinateOutOfBoundsException";
     }
-    if (yCoord>1000.0 || yCoord<0.0) {
+    if (yCoord>1000 || yCoord<0) {
         cerr << "Please make sure Y-Coordinate is between 0.0 and 1000.0";
         throw "CoordinateOutOfBoundsException";
     }
@@ -23,6 +23,6 @@ city::city(std::string name, double xCoord, double yCoord) : name(name) {
 
 std::string city::getName() const { return this->name; }
 
-double city::getX() const { return this->x; }
+int city::getX() const { return this->x; }
 
-double city::getY() const { return this->y; }
+int city::getY() const { return this->y; }
